+++ python/cudf/cudf/_libxx/__init__.py
@@ -5,6 +5,7 @@ import numpy as np
 from . import (
     avro,
     copying,
+    dlpack,
     gpuarrow,
     hash,
     join,
@@ -17,6 +18,7 @@ from . import (
     search,
     sort,
     stream_compaction,
+    strings,
     table,
     transpose,
     unary,