+++ CHANGELOG.md
@@ -116,6 +116,7 @@
 - PR #4233 Porting replace.pyx to use new libcudf APIs
 - PR #4223 Fix a few of the Cython warnings
 - PR #4234 Add BUILD_LEGACY_TESTS cmake option
+- PR #4244 Port nvstrings Substring Gather/Scatter functions to cuDF Python/Cython
 - PR #4251 Add class to docs in `dask-cudf` `derived_from`
 - PR #4261 libxx Cython reorganization
 - PR #4274 Support negative position values in slice_strings